Heavy renaming to get rid of SAM2 (vs.3) biases
Files renamed (under octron/sam_octron/helpers/):
•  sam2_colors.py → octron_colors.py
•  sam2_layer_callback.py → sam_layer_callback.py
•  sam2_layer.py → sam_layer.py
•  sam2_zarr.py → sam_zarr.py

Imports updated in 11 files across octron/:
•  main.py — 3 import lines
•  sam_layer_callback.py — 2 import lines (colors + zarr)
•  sam_layer.py — 2 import lines (zarr + colors lazy import)
•  object_organizer.py — 2 import lines (zarr + colors)
•  yolo_results.py — 2 import lines (zarr + colors)
•  sam2_octron.py, sam2hq_octron.py, sam3_octron.py — relative .sam_zarr import
•  yolo_octron.py, training.py, sam2_results.py — absolute zarr import

•  sam_layer_callback.py: sam2_octron_callbacks → sam_octron_callbacks
•  sam_layer.py: add_sam2_mask_layer → add_sam_mask_layer, add_sam2_shapes_layer → add_sam_shapes_layer, add_sam2_points_layer → add_sam_points_layer
•  main.py: Updated all imports and usages to match the new names
