+++ octron/sam_octron/helpers/sam2_octron.py
@@ -18,7 +18,7 @@ from torch import tensor as torch_tensor
 from skimage.morphology import disk
 
 # Custom Zarr archive class
-from .sam2_zarr import OctoZarr
+from .sam_zarr import OctoZarr
 
 import warnings
 warnings.simplefilter("ignore")
